requirements.txt
django==3.2
psycopg2==2.8.6
pillow==8.1.2
djangorestframework==3.12.4
bcrypt==3.2.0
